api path changes in auth
